# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Agenvas API
  version: 0.1.0
  description: |
    Agenvas 的权威 HTTP 合约。覆盖当前已实现的身份、项目、画布、Agent、Task 与私有素材切片；未实现的 MVP 能力继续按开发清单加入。
    部署者设置 AGENVAS_RECOVERY_MODE=true 后，/api/v1/projects、/api/v1/settings 与 /api/v1/media-templates 下的非 GET/HEAD/OPTIONS 请求统一返回 HTTP 503 ProblemDetail，code=RECOVERY_MODE_READ_ONLY；身份登录与只读核对仍可用。恢复模式同时关闭后台调度，不能把 503 当作业务版本冲突重试。
    服务端为每次 HTTP 请求生成 X-Request-Id 响应头；ProblemDetail.traceId 与它相同。客户端同名请求头不会被信任或回显。
    所有 API 接受 Accept-Language 语言协商（en、zh、ru、ja，地区变体按基础语言匹配；不支持、缺失或无效时回退 zh，q=0 不选用）。响应携带 Content-Language 与 Vary: Accept-Language，包括认证和 CSRF 错误。Problem 的 title、detail、fieldErrors.message 使用协商语言；code、状态、字段路径、ID 和用户内容不随语言变化。语言偏好仅由浏览器保存，无账号配置 API 或数据库迁移。
servers:
  - url: /
paths:
  /api/v1/media-styles:
    get:
      operationId: listMediaStyles
      summary: 图片与视频的全局风格目录，不返回后台注入提示词
      tags: [Settings]
      responses:
        "200":
          description: 包含停用条目以保留既有选择；Cache-Control no-store
          content:
            application/json:
              schema: {type: array, items: {$ref: "#/components/schemas/MediaStyleSummary"}}
        "401": {$ref: "#/components/responses/Unauthenticated"}
  /api/v1/media-styles/{styleId}/thumbnail:
    parameters:
      - name: styleId
        in: path
        required: true
        schema: {type: string, format: uuid}
    get:
      operationId: getMediaStyleThumbnail
      summary: 已登录用户读取风格效果预览
      tags: [Settings]
      responses:
        "200":
          description: 预置 WebP 或实际解码后重新编码的 PNG；Cache-Control no-store
          content:
            image/webp: {schema: {type: string, format: binary}}
            image/png: {schema: {type: string, format: binary}}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "404": {$ref: "#/components/responses/NotFound"}
  /api/v1/settings/media-styles:
    get:
      operationId: listMediaStyleSettings
      summary: 管理员读取风格名称、分类、提示词和配图状态
      tags: [Settings]
      responses:
        "200":
          description: 全局风格设置；Cache-Control no-store
          content:
            application/json:
              schema: {type: array, items: {$ref: "#/components/schemas/MediaStyle"}}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
    post:
      operationId: createMediaStyle
      summary: 管理员手动新增风格；配图随后通过上传接口设置
      tags: [Settings]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateMediaStyleRequest"}
      responses:
        "201":
          description: 风格已创建；Cache-Control no-store
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MediaStyle"}
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "503": {$ref: "#/components/responses/Unavailable"}
  /api/v1/settings/media-styles/{styleId}:
    parameters:
      - name: styleId
        in: path
        required: true
        schema: {type: string, format: uuid}
    put:
      operationId: updateMediaStyle
      summary: 管理员以 CAS 修改或停用风格，已受理任务保持原快照
      tags: [Settings]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/UpdateMediaStyleRequest"}
      responses:
        "200":
          description: 风格已保存；Cache-Control no-store
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MediaStyle"}
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "409": {$ref: "#/components/responses/Conflict"}
        "503": {$ref: "#/components/responses/Unavailable"}
  /api/v1/settings/media-styles/{styleId}/thumbnail:
    parameters:
      - name: styleId
        in: path
        required: true
        schema: {type: string, format: uuid}
    post:
      operationId: uploadMediaStyleThumbnail
      summary: 管理员上传风格配图；禁止外部 URL
      description: 实际 PNG/JPEG/WebP 解码；至多 5 MiB 与 2000 万像素，最长边缩小至 640 像素，去除元数据后重新编码 PNG，保存上限 1 MiB。
      tags: [Settings]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, expectedVersion]
              properties:
                file: {type: string, format: binary}
                expectedVersion: {type: integer, format: int64, minimum: 0}
      responses:
        "200":
          description: 配图已替换，风格版本递增；Cache-Control no-store
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MediaStyle"}
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "409": {$ref: "#/components/responses/Conflict"}
        "503": {$ref: "#/components/responses/Unavailable"}
...
  /api/v1/projects/{projectId}/agents/{agentId}:
    parameters:
      - {$ref: "#/components/parameters/AcceptLanguage"}
      - {$ref: "#/components/parameters/ProjectId"}
      - name: agentId
        in: path
        required: true
        schema: {type: string, format: uuid}
    get:
      operationId: getAgent
      summary: 获取一个 Agent 配置
      tags: [Agents]
      responses:
        "200":
          description: Agent 配置
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Agent"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "404": {$ref: "#/components/responses/NotFound"}
    patch:
      operationId: updateAgent
      summary: 乐观替换 Agent 配置和明确输入绑定
      tags: [Agents]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/UpdateAgentRequest"}
      responses:
        "200":
          description: 更新后的 Agent
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Agent"}
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
...
  /api/v1/projects/{projectId}/agents/{agentId}/skill-binding:
    parameters:
      - {$ref: "#/components/parameters/AcceptLanguage"}
      - name: projectId
        in: path
        required: true
        schema: {type: string, format: uuid}
      - name: agentId
        in: path
        required: true
        schema: {type: string, format: uuid}
    get:
      operationId: getAgentSkillBinding
      tags: [Skills]
      responses:
        "200":
          description: 已鉴权的固定内容或持久操作状态
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AgentSkillBinding"}
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
    put:
      operationId: saveAgentSkillBinding
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: {type: string, minLength: 1, maxLength: 200}
      tags: [Skills]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/SaveAgentSkillBindingRequest"}
      responses:
        "200":
          description: 已鉴权的固定内容或持久操作状态
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AgentSkillBinding"}
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
  /api/v1/projects/{projectId}/agents/{agentId}/skill-installations:
    parameters:
      - {$ref: "#/components/parameters/AcceptLanguage"}
      - name: projectId
        in: path
        required: true
        schema: {type: string, format: uuid}
      - name: agentId
        in: path
        required: true
        schema: {type: string, format: uuid}
    post:
      operationId: installAgentSkill
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: {type: string, minLength: 1, maxLength: 200}
      tags: [Skills]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/InstallAgentSkillRequest"}
      responses:
        "202":
          description: 已鉴权的固定内容或持久操作状态
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SkillInstallation"}
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
  /api/v1/projects/{projectId}/agents/{agentId}/skill-installations/{operationId}:
    parameters:
      - {$ref: "#/components/parameters/AcceptLanguage"}
      - name: projectId
        in: path
        required: true
        schema: {type: string, format: uuid}
      - name: agentId
        in: path
        required: true
        schema: {type: string, format: uuid}
      - name: operationId
        in: path
        required: true
        schema: {type: string, format: uuid}
    get:
      operationId: getAgentSkillInstallation
      tags: [Skills]
      responses:
        "200":
          description: 已鉴权的固定内容或持久操作状态
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SkillInstallation"}
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
components:
  schemas:
    MediaStyleSummary:
      type: object
      additionalProperties: false
      required: [id, name, category, enabled, version, thumbnailUrl, builtIn]
      properties:
        id: {type: string, format: uuid}
        name: {type: string, minLength: 1, maxLength: 80}
        category: {type: string, minLength: 1, maxLength: 40}
        enabled: {type: boolean}
        version: {type: integer, format: int64, minimum: 0}
        thumbnailUrl: {type: [string, "null"], description: Authenticated application URL; null until a custom preview is uploaded.}
        builtIn: {type: boolean}
    MediaStyle:
      type: object
      additionalProperties: false
      required: [id, name, category, enabled, version, thumbnailUrl, builtIn, promptSuffix]
      properties:
        id: {type: string, format: uuid}
        name: {type: string, minLength: 1, maxLength: 80}
        category: {type: string, minLength: 1, maxLength: 40}
        enabled: {type: boolean}
        version: {type: integer, format: int64, minimum: 0}
        thumbnailUrl: {type: [string, "null"]}
        builtIn: {type: boolean}
        promptSuffix: {type: string, minLength: 1, maxLength: 4000}
    CreateMediaStyleRequest:
      type: object
      additionalProperties: false
      required: [name, category, promptSuffix, enabled]
      properties:
        name: {type: string, minLength: 1, maxLength: 80}
        category: {type: string, minLength: 1, maxLength: 40}
        promptSuffix: {type: string, minLength: 1, maxLength: 4000}
        enabled: {type: boolean}
    UpdateMediaStyleRequest:
      type: object
      additionalProperties: false
      required: [expectedVersion, name, category, promptSuffix, enabled]
      properties:
        expectedVersion: {type: integer, format: int64, minimum: 0}
        name: {type: string, minLength: 1, maxLength: 80}
        category: {type: string, minLength: 1, maxLength: 40}
        promptSuffix: {type: string, minLength: 1, maxLength: 4000}
        enabled: {type: boolean}
...
    AgentBindingRequest:
      type: object
      additionalProperties: false
      required: [artifactId, selectedVersionId]
      properties:
        artifactId: {type: string, format: uuid}
        selectedVersionId: {type: string, format: uuid}
...
    UpdateAgentRequest:
      type: object
      additionalProperties: false
      required: [expectedVersion, name, instruction, bindings]
      properties:
        expectedVersion: {type: integer, format: int64, minimum: 0}
        name: {type: string, minLength: 1, maxLength: 120}
        instruction: {type: string, minLength: 1, maxLength: 8000}
        bindings:
          type: array
          maxItems: 40
          items: {$ref: "#/components/schemas/AgentBindingRequest"}
    AgentBinding:
      type: object
      additionalProperties: false
      required: [id, artifactId, selectedVersionId]
      properties:
        id: {type: string, format: uuid}
        artifactId: {type: string, format: uuid}
        selectedVersionId: {type: string, format: uuid}
    Agent:
      type: object
      additionalProperties: false
      required: [id, projectId, profileKey, profileVersion, name, instruction, outputGroupId, version, createdAt, updatedAt, bindings]
      properties:
        id: {type: string, format: uuid}
        projectId: {type: string, format: uuid}
        profileKey: {type: string, const: creator}
        profileVersion: {type: integer, minimum: 1}
        name: {type: string}
        instruction: {type: string}
        outputGroupId: {type: string, format: uuid}
        version: {type: integer, format: int64, minimum: 0}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
        bindings:
          type: array
          items: {$ref: "#/components/schemas/AgentBinding"}
...
    Problem:
      type: object
      additionalProperties: true
      required: [type, title, status, code, traceId, retryable]
      properties:
        type: {type: string, format: uri}
        title: {type: string}
        status: {type: integer, minimum: 400, maximum: 599}
        code: {type: string}
        detail: {type: string}
        instance: {type: string, format: uri-reference}
        traceId: {type: string}
        retryable: {type: boolean}
        fieldErrors:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [field, message]
            properties:
              field: {type: string}
              message: {type: string}
...
    AgentSkillBinding:
      type: object
      additionalProperties: false
      required: [agentVersion, skillId, skillVersionId]
      properties:
        agentVersion: {type: integer, format: int64, minimum: 0}
        skillId: {type: [string, "null"], format: uuid}
        skillVersionId: {type: [string, "null"], format: uuid}
    SaveAgentSkillBindingRequest:
      type: object
      additionalProperties: false
      required: [expectedAgentVersion, skillId, skillVersionId]
      properties:
        expectedAgentVersion: {type: integer, format: int64, minimum: 0}
        skillId: {type: [string, "null"], format: uuid}
        skillVersionId: {type: [string, "null"], format: uuid}
    InstallAgentSkillRequest:
      type: object
      additionalProperties: false
      required: [skillId, skillVersionId]
      properties:
        skillId: {type: string, format: uuid}
        skillVersionId: {type: string, format: uuid}
    SkillInstallation:
      type: object
      additionalProperties: false
      required: [id, status, skillId, skillVersionId, errorCode, errorDetail]
      properties:
        id: {type: string, format: uuid}
        status: {type: string, enum: [ACCEPTED, PREPARING, SUCCEEDED, FAILED, CLEANING]}
        skillId: {type: string, format: uuid}
        skillVersionId: {type: string, format: uuid}
        errorCode: {type: [string, "null"]}
        errorDetail: {type: [string, "null"]}
...
  responses:
    ValidationError:
      description: 请求参数无效
      headers:
        Content-Language: {$ref: "#/components/headers/ContentLanguage"}
        Vary: {$ref: "#/components/headers/VaryLanguage"}
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
    Unauthenticated:
      description: 未登录或凭据不正确
      headers:
        Content-Language: {$ref: "#/components/headers/ContentLanguage"}
        Vary: {$ref: "#/components/headers/VaryLanguage"}
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
    Forbidden:
      description: CSRF、初始化凭据或权限校验失败
      headers:
        Content-Language: {$ref: "#/components/headers/ContentLanguage"}
        Vary: {$ref: "#/components/headers/VaryLanguage"}
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
    Conflict:
      description: 初始化或乐观版本冲突
      headers:
        Content-Language: {$ref: "#/components/headers/ContentLanguage"}
        Vary: {$ref: "#/components/headers/VaryLanguage"}
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
    NotFound:
      description: 资源不存在或当前用户无权访问
      headers:
        Content-Language: {$ref: "#/components/headers/ContentLanguage"}
        Vary: {$ref: "#/components/headers/VaryLanguage"}
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
...
    Unavailable:
      description: 部署凭证加密主密钥尚未配置
      headers:
        Content-Language: {$ref: "#/components/headers/ContentLanguage"}
        Vary: {$ref: "#/components/headers/VaryLanguage"}
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
...
  headers:
    ContentLanguage:
      description: 协商后的界面语言；不改变用户内容、枚举或错误码。
      schema: {type: string, enum: [en, zh, ru, ja]}
    VaryLanguage:
      description: 包含 Accept-Language，避免不同语言的响应共用缓存。
      schema: {type: string}
  parameters:
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      description: 支持 en、zh、ru、ja 及其地区变体，按质量权重选择；缺失、不支持或无效时回退 zh。
      schema: {type: string}
      example: ru-RU, en;q=0.8
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: {type: string, format: uuid}
...